Improve responsiveness of quilt composition to MUIBar ChartScale dialog.

+++ include/chcanv.h
@@ -473,7 +473,7 @@ public:
       void ToggleCanvasQuiltMode( void );
       
       wxString GetScaleText(){ return m_scaleText; }
-      int GetScaleValue(){ return m_scaleValue; }
+      double GetScaleValue(){ return m_scaleValue; }
       
       bool        m_b_paint_enable;
 
@@ -550,7 +550,7 @@ private:
       int         cursor_region;
       bool        m_bTCupdate;
       wxString    m_scaleText;
-      int         m_scaleValue;
+      double      m_scaleValue;
       bool        m_bShowScaleInStatusBar;
       wxRect      bbRect;
